+++ src/Row.php
@@ -249,14 +249,7 @@ class Row implements JsonSerializable
             }
 
             if (empty($this->id)) {
-                $values = array_filter(
-                    $this->toArray(),
-                    function ($value) {
-                        return $value !== null;
-                    }
-                );
-
-                $this->id = $this->table->insert($values)->run();
+                $this->id = $this->table->insert($this->toArray())->run();
             } elseif (empty($this->changes)) {
                 return $this;
             } else {